+++ flake8_docstrings_complete/__init__.py
@@ -18,6 +18,11 @@ DOCSTR_MISSING_MSG = (
     f"{DOCSTR_MISSING_CODE} docstring should be defined for a function/ method/ class"
     f"{MORE_INFO_BASE}{DOCSTR_MISSING_CODE.lower()}"
 )
+MAGIC_METHOD_DOCSTR_MISSING_CODE = f"{ERROR_CODE_PREFIX}011"
+MAGIC_METHOD_DOCSTR_MISSING_MSG = (
+    f"{MAGIC_METHOD_DOCSTR_MISSING_CODE} docstring should be defined for a magic method"
+    f"{MORE_INFO_BASE}{MAGIC_METHOD_DOCSTR_MISSING_CODE.lower()}"
+)
 RETURNS_SECTION_NOT_IN_DOCSTR_CODE = f"{ERROR_CODE_PREFIX}030"
 RETURNS_SECTION_NOT_IN_DOCSTR_MSG = (
     f"{RETURNS_SECTION_NOT_IN_DOCSTR_CODE} function/ method that returns a value should have the "
@@ -360,9 +365,14 @@ class Visitor(ast.NodeVisitor):
         if not self._skip_function(node=node):
             # Check docstring is defined
             if ast.get_docstring(node) is None:
+                docstr_missing_msg = (
+                    MAGIC_METHOD_DOCSTR_MISSING_MSG
+                    if node.name.startswith("__") and node.name.endswith("__")
+                    else DOCSTR_MISSING_MSG
+                )
                 self.problems.append(
                     types_.Problem(
-                        lineno=node.lineno, col_offset=node.col_offset, msg=DOCSTR_MISSING_MSG
+                        lineno=node.lineno, col_offset=node.col_offset, msg=docstr_missing_msg
                     )
                 )