+++ tests/unit/test___init__.py
@@ -6,6 +6,7 @@ import pytest
 
 from flake8_docstrings_complete import (
     DOCSTR_MISSING_MSG,
+    MAGIC_METHOD_DOCSTR_MISSING_MSG,
     MULT_RETURNS_SECTIONS_IN_DOCSTR_MSG,
     MULT_YIELDS_SECTIONS_IN_DOCSTR_MSG,
     RETURNS_SECTION_IN_DOCSTR_MSG,
@@ -28,6 +29,14 @@ def function_1():
 """,
             (f"2:0 {DOCSTR_MISSING_MSG}",),
             id="function docstring missing return",
+        ),
+        pytest.param(
+            """
+def __function_1__():
+    return
+""",
+            (f"2:0 {MAGIC_METHOD_DOCSTR_MISSING_MSG}",),
+            id="magic function docstring missing return",
         ),
         pytest.param(
             """
@@ -487,6 +496,16 @@ class Class1:
 ''',
             (f"4:4 {DOCSTR_MISSING_MSG}",),
             id="method docstring missing return",
+        ),
+        pytest.param(
+            '''
+class Class1:
+    """Docstring."""
+    def __init__(self):
+        return
+''',
+            (f"4:4 {MAGIC_METHOD_DOCSTR_MISSING_MSG}",),
+            id="magic method docstring missing return",
         ),
         pytest.param(
             '''